+++ service/internal/cloudconfig/template/etcd3_attach_dep_service.go
@@ -9,7 +9,7 @@ Before=k8s-kubelet.service
 
 [Service]
 # image is from https://github.com/giantswarm/aws-attach-etcd-dep
-Environment="IMAGE={{ .RegistryDomain }}/giantswarm/aws-attach-etcd-dep:27182d5b07d824576dd47869e5f921c1341fb194"
+Environment="IMAGE={{ if ne .RegistryDomain "" }}{{ .RegistryDomain }}/{{ end }}/giantswarm/aws-attach-etcd-dep:27182d5b07d824576dd47869e5f921c1341fb194"
 Environment="NAME=%p.service"
 Type=oneshot
 RemainAfterExit=yes